[AI4DSOC][Security Solution] Extract takeActions logic outside the detections grouping alerts table
