+++ x-pack/solutions/security/plugins/security_solution/public/detections/components/alerts_table/alerts_grouping.tsx
@@ -25,7 +25,6 @@ import { useIsExperimentalFeatureEnabled } from '../../../common/hooks/use_exper
 import { groupIdSelector } from '../../../common/store/grouping/selectors';
 import { useDeepEqualSelector } from '../../../common/hooks/use_selector';
 import { updateGroups } from '../../../common/store/grouping/actions';
-import type { Status } from '../../../../common/api/detection_engine';
 import { defaultUnit } from '../../../common/components/toolbar/unit';
 import { useSourcererDataView } from '../../../sourcerer/containers';
 import type { RunTimeMappings } from '../../../sourcerer/store/model';
@@ -59,7 +58,6 @@ export interface AlertsTableComponentProps {
      */
     renderer: GetGroupStats<AlertsGroupingAggregation>;
   };
-  currentAlertStatusFilterValue?: Status[];
   defaultFilters?: Filter[];
   /**
    * Default values to display in the group selection dropdown.
@@ -69,8 +67,16 @@ export interface AlertsTableComponentProps {
   from: string;
   globalFilters: Filter[];
   globalQuery: Query;
-  hasIndexMaintenance: boolean;
-  hasIndexWrite: boolean;
+  /**
+   * Allows to customize the content of the Take actions button rendered at the group level.
+   * If no value is provided, the Take actins button is not displayed.
+   */
+  groupTakeActionItems?: (data: {
+    query?: string;
+    tableId: string;
+    groupNumber: number;
+    selectedGroup: string;
+  }) => JSX.Element[];
   loading: boolean;
   renderChildComponent: (groupingFilters: Filter[]) => React.ReactElement;
   runtimeMappings: RunTimeMappings;
@@ -341,6 +347,7 @@ const GroupedAlertsTableComponent: React.FC<AlertsTableComponentProps> = (props)
           getGrouping={getGrouping}
           groupingLevel={level}
           groupStatsAggregations={groupStatusAggregations}
+          groupTakeActionItems={props.groupTakeActionItems}
           onGroupClose={() => resetGroupChildrenPagination(level)}
           pageIndex={pageIndex[level] ?? DEFAULT_PAGE_INDEX}
           pageSize={pageSize[level] ?? DEFAULT_PAGE_SIZE}
