+++ x-pack/solutions/security/plugins/security_solution/public/detections/components/alerts_table/grouping_settings/index.tsx
@@ -9,5 +9,4 @@ export * from './default_grouping_options';
 export * from './default_group_stats_aggregations';
 export * from './default_group_stats_renderers';
 export * from './default_group_title_renderers';
-export * from './group_take_action_items';
 export * from './query_builder';
